+++ release_notes.md
@@ -10,7 +10,9 @@
 
 ### Enhancements:
 
-* Lorem ipsum.
+* `WriteTransaction::has_table()` and `WriteTransaction::rollback()` were
+  added. Previously, only implicit rollback was possible with
+  `WriteTransaction`.
 
 -----------